backend/pyproject.toml
[tool.poetry]
name = "asila-backend"
version = "0.1.0"
description = "Backend services for Asila WhatsApp public info system"
authors = ["Asila Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
httpx = "^0.27.0"
python-multipart = "^0.0.9"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
sqlalchemy = { version = "^2.0.32", extras = ["asyncio"] }
asyncpg = "^0.29.0"
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
ruff = "^0.5.6"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
